Add files via upload
Abstract—Landslide susceptibility mapping and modelling
that deal directly with the forecast of landslides
have been researched in the past. However, this study
presents landslide trigger classification as a novel approach
in determining areas that are potentially prone to
landslides. Random forest (RF) is used on the multi-class
category on the landslide dataset. Average performance
was observed and when the SMOTE sampling was
done to combat class imbalance before implementing
RF, the performance was observed to be better than
without. Kernelised SVM with four different kernels
were tried out with 25 combinations of hyper-parameters.
The kernels help define the hyperplane that is used for
the classification problem. The hyper-parameters that
brought out the optimal performance is with gamma
value of 0.1 and cost function value of 100. The radial
kernel outperformed the other kernels with a better
recorded accuracy. Landslide trigger classification are
used in drawing up warning zones prior to the occurrence
of landslides by tracking the landslide triggers.